+++ tests/testthat/test_6_qm_summarize.R
@@ -25,57 +25,79 @@ test_tbl <- as_tibble(data.frame(
 # test inputs ------------------------------------------------
 
 # test missing ref parameter
-expect_error(qm_summarize(key = TRACTCE, clusters = clusters, category = "positive"),
+expect_error(qm_summarize(key = TRACTCE, clusters = clusters, category = "positive", count = "clusters"),
              "A reference, consisting of a simple features object, must be specified.")
 
 # test no sf ref parameter
-expect_error(qm_summarize(ref = test_tbl, key = TRACTCE, clusters = clusters, category = "positive"),
+expect_error(qm_summarize(ref = test_tbl, key = TRACTCE, clusters = clusters, category = "positive",
+                          count = "clusters"),
              "The reference object must be a simple features object.")
 
 # test missing key parameter
-expect_error(qm_summarize(ref = test_sf, clusters = clusters, category = "positive"),
+expect_error(qm_summarize(ref = test_sf, clusters = clusters, category = "positive", count = "clusters"),
              "A key identification variable must be specified.")
 
 # test incorrect key parameter
-expect_error(qm_summarize(ref = test_sf, key = "ham", clusters = clusters, category = "positive"),
+expect_error(qm_summarize(ref = test_sf, key = "ham", clusters = clusters, category = "positive",
+                          count = "clusters"),
              "The specified key ham cannot be found in the reference data.")
-expect_error(qm_summarize(ref = test_sf, key = ham, clusters = clusters, category = "positive"),
+expect_error(qm_summarize(ref = test_sf, key = ham, clusters = clusters, category = "positive",
+                          count = "clusters"),
              "The specified key ham cannot be found in the reference data.")
-expect_error(qm_summarize(ref = test_sf, key = "TRACTCE", clusters = clusters2, category = "positive"),
+expect_error(qm_summarize(ref = test_sf, key = "TRACTCE", clusters = clusters2, category = "positive",
+                          count = "clusters"),
              "The specified key TRACTCE cannot be found in the clusters data.")
-expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters2, category = "positive"),
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters2, category = "positive",
+                          count = "clusters"),
              "The specified key TRACTCE cannot be found in the clusters data.")
 
 # test missing clusters parameter
-expect_error(qm_summarize(ref = test_sf, key = TRACTCE, category = "postitive"),
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, category = "postitive",
+                          count = "clusters"),
              "A data set containing map clusters must be specified.")
 
 # test non qm_cluster object input
-expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = test_tbl, category = "positive"),
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = test_tbl, category = "positive",
+                          count = "clusters"),
              "The object test_tbl is not class qm_cluster. The cluster object should be created with qm_combine().", fixed = TRUE)
 
 # test missing category parameter
-expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters),
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, count = "clusters"),
              "A category from the cluster object must be specified.")
 
 # test incorrect category parameter
-expect_error(qm_summarize(ref = test_sf, key = "TRACTCE", clusters = clusters, category = "ham"),
+expect_error(qm_summarize(ref = test_sf, key = "TRACTCE", clusters = clusters, category = "ham",
+                          count = "clusters"),
              "The specified category ham cannot be found in the clusters data.")
-expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = ham),
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = ham,
+                          count = "clusters"),
              "The specified category ham cannot be found in the clusters data.")
 
+# test missing count parameter
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "postitive"),
+             "A method for producing counts, either 'clusters' or 'respondents', must be specified.")
+
+# test incorrect count parameter
+expect_error(qm_summarize(ref = test_sf, key = "TRACTCE", clusters = clusters, category = "postitive",
+                          count = "ham"),
+             "Counts only accepts 'clusters' or 'respondents' as arguments.")
+
 # test for incorrect geometry parameter
-expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = positive, geometry = "TRUE"),
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = positive,
+                          count = "clusters", geometry = "TRUE"),
              "The geometry parameter only accepts TRUE or FALSE as arguments.")
 
 # test for incorrect use.na parameter
-expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = positive, use.na = "TRUE"),
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = positive,
+                          count = "clusters", use.na = "TRUE"),
              "The use.na parameter only accepts TRUE or FALSE as arguments.")
 
 # test results ------------------------------------------------
 
-resultV1 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive", geometry = TRUE, use.na = FALSE)
-resultV1b <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive")
+resultV1 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive",
+                         count = "clusters", geometry = TRUE, use.na = FALSE)
+resultV1b <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive",
+                          count = "clusters")
 
 nrowV1 <- 106
 posV1 <- 0.05660377
@@ -89,7 +111,8 @@ test_that("result object 1 has expected characteristics", {
   expect_equal(resultV1, resultV1b)
 })
 
-resultV2 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive", geometry = TRUE, use.na = TRUE)
+resultV2 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive",
+                         count = "clusters", geometry = TRUE, use.na = TRUE)
 
 nrowV2 <- 106
 posV2 <- 1.2
@@ -102,7 +125,8 @@ test_that("result object 2 has expected characteristics", {
   expect_equal("sf", classListElementV2)
 })
 
-resultV3 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive", geometry = FALSE, use.na = FALSE)
+resultV3 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive",
+                         count = "clusters", geometry = FALSE, use.na = FALSE)
 
 nrowV3 <- 5
 posV3 <- 1.2