+++ NEWS.md
@@ -1,3 +1,9 @@
+# qualmap 0.2.0.9000
+
+* Add `qm_verify()` as a means for verifying data data previously saved to disk prior to processing them with `qm_summarize()`
+* Add second approach to producing counts using `qm_summarize()` that returns counts of participants rather than counts of clusters associated with each feature
+* Remove the inclusion of the `COUNT` from what is returned with `qm_create()`
+
 # qualmap 0.1.1
 
 * CRAN release version